backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-anthropic
pydantic
python-multipart
python-dotenv
